IGNITE-19676 JMX Invoker
