--- modules/core/src/main/java/org/apache/ignite/internal/management/jmx/JmxCommandRegistryInvokerPluginProvider.java
@@ -1,110 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.ignite.internal.management.jmx;
-
-import java.io.Serializable;
-import java.util.UUID;
-import org.apache.ignite.IgniteCheckedException;
-import org.apache.ignite.cluster.ClusterNode;
-import org.apache.ignite.internal.IgniteVersionUtils;
-import org.apache.ignite.plugin.CachePluginContext;
-import org.apache.ignite.plugin.CachePluginProvider;
-import org.apache.ignite.plugin.ExtensionRegistry;
-import org.apache.ignite.plugin.IgnitePlugin;
-import org.apache.ignite.plugin.PluginContext;
-import org.apache.ignite.plugin.PluginProvider;
-import org.apache.ignite.plugin.PluginValidationException;
-import org.jetbrains.annotations.Nullable;
-
-/**
- *
- */
-public class JmxCommandRegistryInvokerPluginProvider implements PluginProvider {
-    /** */
-    private JmxComandRegistryInvoker jmxInvoker = new JmxComandRegistryInvoker();
-
-    /** {@inheritDoc} */
-    @Override public String name() {
-        return "JMX invoker of CommandRegistry";
-    }
-
-    /** {@inheritDoc} */
-    @Override public String version() {
-        return IgniteVersionUtils.VER_STR;
-    }
-
-    /** {@inheritDoc} */
-    @Override public String copyright() {
-        return IgniteVersionUtils.COPYRIGHT;
-    }
-
-    /** {@inheritDoc} */
-    @Override public IgnitePlugin plugin() {
-        return jmxInvoker;
-    }
-
-    /** {@inheritDoc} */
-    @Override public void initExtensions(PluginContext ctx, ExtensionRegistry registry) throws IgniteCheckedException {
-        jmxInvoker.context(ctx);
-    }
-
-    /** {@inheritDoc} */
-    @Override public CachePluginProvider createCacheProvider(CachePluginContext ctx) {
-        return null;
-    }
-
-    /** {@inheritDoc} */
-    @Override public void onIgniteStart() throws IgniteCheckedException {
-        jmxInvoker.onIgniteStart();
-    }
-
-    /** {@inheritDoc} */
-    @Override public void onIgniteStop(boolean cancel) {
-        jmxInvoker.onIgniteStop();
-    }
-
-    /** {@inheritDoc} */
-    @Override public @Nullable Serializable provideDiscoveryData(UUID nodeId) {
-        return null;
-    }
-
-    /** {@inheritDoc} */
-    @Override public void receiveDiscoveryData(UUID nodeId, Serializable data) {
-        // No-op.
-    }
-
-    /** {@inheritDoc} */
-    @Override public void validateNewNode(ClusterNode node) throws PluginValidationException {
-        // No-op.
-    }
-
-    /** {@inheritDoc} */
-    @Override public Object createComponent(PluginContext ctx, Class cls) {
-        return null;
-    }
-
-    /** {@inheritDoc} */
-    @Override public void start(PluginContext ctx) throws IgniteCheckedException {
-        // No-op.
-    }
-
-    /** {@inheritDoc} */
-    @Override public void stop(boolean cancel) throws IgniteCheckedException {
-        // No-op.
-    }
-}
